Move danger utils to another file
